requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.7.0
python-dotenv>=1.0.1
requests>=2.32.0
scikit-learn>=1.6.0
numpy>=2.1.0
faiss-cpu>=1.8.0
